added ingredients list
With the help of stack overflow research, I feel vindicated to know that the drinks api is poorly designed in the opinion of many devs, as having each ingredient separated rather than in an array proved to be an unseen challenge by me. I gained exposure to some interesting approaches to this problem, and saw how clever one can get when you have to refactor data that is served poorly. Drinks up! 🍻🥂